+++ darwin/photo_manager/Sources/photo_manager/core/PMManager.h
@@ -1,6 +1,5 @@
 #import "PMFileHelper.h"
 #import "PMImport.h"
-#import <CoreLocation/CoreLocation.h>
 #import <Foundation/Foundation.h>
 #import <Photos/Photos.h>